requirements.txt
# Core Dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
httpx>=0.23.0
requests>=2.28.0
python-dotenv>=1.0.0

# Google AI / Gemini
google-generativeai>=0.3.0

# Database
aiosqlite>=0.19.0

# Vector Databases
chromadb>=0.4.0
pinecone-client>=2.2.0

# LangFlow
langflow>=0.6.0

# Search APIs
google-search-results>=2.4.2

# Async Support
aiohttp>=3.8.0
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # Parallel test execution
respx>=0.20.0  # HTTP mocking for httpx
freezegun>=1.2.0  # Time mocking
faker>=19.0.0  # Test data generation

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Utilities
tenacity>=8.2.0
tiktoken>=0.5.0

# Monitoring and Observability
prometheus-client>=0.17.0
structlog>=23.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-sqlite3>=0.41b0
psutil>=5.9.0  # For system metrics

# Security
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.6  # Form data parsing
redis>=5.0.0  # Rate limiting backend (optional)

# Production Server
uvloop>=0.17.0  # High-performance event loop
gunicorn>=21.2.0  # WSGI server (alternative)

